+++ examples/polkadot/assetHub/foreignAssetTransfers/local/assetHubKSMTransfer.ts
@@ -8,27 +8,26 @@ import { TxResult } from '../../../../../src/types';
 import { GREEN, PURPLE, RESET } from '../../../../colors';
 
 /**
- * In this example we are creating a `polkadotXcm` pallet `limitedReserveTransferAssets` call to send MOVR (foreign asset with location `{"parents":"1","interior":{"X2":[{"Parachain":"2125"},{"GeneralIndex":"0"}]}}`)
- * from a Kusama Asset Hub (System Parachain) account
- * to a Bifrost (ParaChain) account, where the `xcmVersion` is set to 3 and no `weightLimit` option is provided declaring that
- * the tx will allow unlimited weight to be used for fees.
+ * In this example we are creating a `foreignAssets` pallet `transfer` call to send 1 KSM (foreign asset with location {"parents":"2","interior":{"X1":{"GlobalConsensus":"Kusama"}}}`)
+ * from a Polkadot Asset Hub (System Parachain) account
+ * to a Polkadot Asset Hub (System Parachain) account.
  *
  * NOTE: To specify the amount of weight for the tx to use provide a `weightLimit` option containing desired values for `refTime` and `proofSize`.
  */
 const main = async () => {
-	const { api, specName, safeXcmVersion } = await constructApiPromise('wss://kusama-asset-hub-rpc.polkadot.io');
+	const { api, specName, safeXcmVersion } = await constructApiPromise('wss://polkadot-asset-hub-rpc.polkadot.io');
 	const assetApi = new AssetTransferApi(api, specName, safeXcmVersion);
 
 	let callInfo: TxResult<'call'>;
 	try {
 		callInfo = await assetApi.createTransferTransaction(
-			'2001', // Note: Parachain ID 2001 (Bifrost) is different than the asset location's `Parachain` Id, making this a `limitedReserveTransferAssets` call
+			'1000', // NOTE: The destination id is `1000` and matches the origin chain making this a local transfer
 			'5EWNeodpcQ6iYibJ3jmWVe85nsok1EDG8Kk3aFg8ZzpfY1qX',
-			['{"parents":"1","interior":{"X2":[{"Parachain":"2023"},{"PalletInstance":"10"}]}}'],
+			['{"parents":"2","interior":{"X1":{"GlobalConsensus":"Kusama"}}}'],
 			['1000000000000'],
 			{
 				format: 'call',
-				xcmVersion: 3,
+				xcmVersion: safeXcmVersion, // Note: GlobalConsensus junctions require XCM V3 or higher
 			},
 		);