+++ src/createXcmTypes/util/getAssetId.ts
@@ -81,7 +81,7 @@ export const getAssetId = async (
 		if (multiLocationIsInRegistry) {
 			assetId = asset;
 		} else {
-			const isValidForeignAsset = await foreignAssetsMultiLocationExists(api, registry, asset, xcmVersion);
+			const isValidForeignAsset = await foreignAssetsMultiLocationExists(api, registry, asset);
 
 			if (!isValidForeignAsset) {
 				throw new BaseError(`MultiLocation ${asset} not found`, BaseErrorsEnum.AssetNotFound);
@@ -112,7 +112,7 @@ export const getAssetId = async (
 			}
 		} else {
 			throw new BaseError(
-				`assetId ${asset} is not a valid symbol or integer asset id for ${specName}`,
+				`assetId ${asset} is not a valid symbol, integer asset id or location for ${specName}`,
 				BaseErrorsEnum.InvalidAsset,
 			);
 		}