+++ src/errors/checkXcmTxInputs.ts
@@ -4,11 +4,19 @@ import type { ApiPromise } from '@polkadot/api';
 import { isHex } from '@polkadot/util';
 import { isEthereumAddress } from '@polkadot/util-crypto';
 
-import { MAX_ASSETS_FOR_TRANSFER, RELAY_CHAIN_IDS } from '../consts';
+import {
+	MAX_ASSETS_FOR_TRANSFER,
+	RELAY_CHAIN_IDS,
+	RELAY_CHAINS_NATIVE_ASSET_LOCATION,
+	SYSTEM_AND_PARACHAINS_RELAY_ASSET_LOCATION,
+} from '../consts';
 import { XcmPalletName } from '../createXcmCalls/util/establishXcmPallet';
+import { assetIdIsLocation } from '../createXcmTypes/util/assetIdIsLocation';
 import { foreignAssetMultiLocationIsInCacheOrRegistry } from '../createXcmTypes/util/foreignAssetMultiLocationIsInCacheOrRegistry';
 import { foreignAssetsMultiLocationExists } from '../createXcmTypes/util/foreignAssetsMultiLocationExists';
+import { getGlobalConsensusSystemName } from '../createXcmTypes/util/getGlobalConsensusSystemName';
 import { isParachainPrimaryNativeAsset } from '../createXcmTypes/util/isParachainPrimaryNativeAsset';
+import { isRelayNativeAsset } from '../createXcmTypes/util/isRelayNativeAsset';
 import { multiLocationAssetIsParachainsNativeAsset } from '../createXcmTypes/util/multiLocationAssetIsParachainsNativeAsset';
 import { Registry } from '../registry';
 import type { ChainInfo, ChainInfoKeys } from '../registry/types';
@@ -238,7 +246,7 @@ export const checkMultiLocationsContainOnlyNativeOrForeignAssetsOfDestChain = (
  * @param assetId
  * @param relayChainInfo
  */
-const checkRelayToSystemAssetId = (assetId: string, relayChainInfo: ChainInfo<ChainInfoKeys>) => {
+const checkRelayToSystemAssetId = (assetId: string, relayChainInfo: ChainInfo<ChainInfoKeys>, registry: Registry) => {
 	const relayChainId = RELAY_CHAIN_IDS[0];
 	const relayChain = relayChainInfo[relayChainId];
 	const relayChainNativeAsset = relayChain.tokens[0];
@@ -246,20 +254,13 @@ const checkRelayToSystemAssetId = (assetId: string, relayChainInfo: ChainInfo<Ch
 	// relay chain can only send its native asset
 	// ensure the asset being sent is the native asset of the relay chain
 	// no need to check if id is a number, if it is, it fails the check by default
-	let assetIsRelayChainNativeAsset = false;
 
-	// if an empty string is passed, treat it as the native relay asset
-	if (assetId === '') {
-		assetIsRelayChainNativeAsset = true;
-	}
-
-	if (relayChainNativeAsset.toLowerCase() === assetId.toLowerCase()) {
-		assetIsRelayChainNativeAsset = true;
-	}
+	// ensure assetId is relay chain's native token
+	const assetIsRelayChainNativeAsset = isRelayNativeAsset(registry, assetId);
 
 	if (!assetIsRelayChainNativeAsset) {
 		throw new BaseError(
-			`(RelayToSystem) asset ${assetId} is not ${relayChain.specName}'s native asset. Expected ${relayChainNativeAsset}`,
+			`(RelayToSystem) assetId ${assetId} is not the native asset of ${relayChain.specName} relay chain. Expected an assetId of ${relayChainNativeAsset} or asset location ${RELAY_CHAINS_NATIVE_ASSET_LOCATION}`,
 			BaseErrorsEnum.InvalidAsset,
 		);
 	}
@@ -294,7 +295,7 @@ const checkRelayToParaAssetId = (assetId: string, relayChainInfo: ChainInfo<Chai
 
 	if (!assetIsRelayChainNativeAsset) {
 		throw new BaseError(
-			`(RelayToPara) asset ${assetId} is not ${relayChain.specName}'s native asset. Expected ${relayChainNativeAsset}`,
+			`(RelayToPara) assetId ${assetId} is not the native asset of ${relayChain.specName} relay chain. Expected an assetId of ${relayChainNativeAsset} or asset location ${RELAY_CHAINS_NATIVE_ASSET_LOCATION}`,
 			BaseErrorsEnum.InvalidAsset,
 		);
 	}
@@ -305,28 +306,17 @@ const checkRelayToParaAssetId = (assetId: string, relayChainInfo: ChainInfo<Chai
  * @param assetId
  * @param relayChainInfo
  */
-const checkSystemToRelayAssetId = (assetId: string, relayChainInfo: ChainInfo<ChainInfoKeys>) => {
+const checkSystemToRelayAssetId = (assetId: string, relayChainInfo: ChainInfo<ChainInfoKeys>, registry: Registry) => {
 	const relayChainId = RELAY_CHAIN_IDS[0];
 	const relayChain = relayChainInfo[relayChainId];
 	const relayChainNativeAsset = relayChain.tokens[0];
 
 	// ensure assetId is relay chain's native token
-	let matchedRelayChainNativeToken = false;
-
-	// if an empty string is passed, treat it as the native relay asset
-	if (assetId === '') {
-		matchedRelayChainNativeToken = true;
-	}
-
-	if (typeof assetId === 'string') {
-		if (relayChainNativeAsset.toLowerCase() === assetId.toLowerCase()) {
-			matchedRelayChainNativeToken = true;
-		}
-	}
+	const matchedRelayChainNativeToken = isRelayNativeAsset(registry, assetId);
 
 	if (!matchedRelayChainNativeToken) {
 		throw new BaseError(
-			`(SystemToRelay) assetId ${assetId} not native to ${relayChain.specName}. Expected ${relayChainNativeAsset}`,
+			`(SystemToRelay) assetId ${assetId} is not the native asset of ${relayChain.specName} relay chain. Expected an assetId of ${relayChainNativeAsset} or asset location ${SYSTEM_AND_PARACHAINS_RELAY_ASSET_LOCATION}`,
 			BaseErrorsEnum.InvalidAsset,
 		);
 	}
@@ -396,13 +386,18 @@ const checkSystemAssets = async (
 	isLiquidTokenTransfer?: boolean,
 ) => {
 	const currentChainId = registry.lookupChainIdBySpecName(specName);
-
-	if (isForeignAssetsTransfer) {
+	if (isForeignAssetsTransfer && assetIdIsLocation(assetId)) {
 		// check that the asset id is a valid multilocation
 		const multiLocationIsInRegistry = foreignAssetMultiLocationIsInCacheOrRegistry(assetId, registry, xcmVersion);
 
 		if (!multiLocationIsInRegistry) {
-			const isValidForeignAsset = await foreignAssetsMultiLocationExists(api, registry, assetId, xcmVersion);
+			const isValidForeignAsset = await foreignAssetsMultiLocationExists(api, registry, assetId);
+
+			if (!isValidForeignAsset) {
+				if (isRelayNativeAsset(registry, assetId)) {
+					return;
+				}
+			}
 
 			if (!isValidForeignAsset) {
 				throw new BaseError(`MultiLocation ${assetId} not found`, BaseErrorsEnum.AssetNotFound);
@@ -445,7 +440,7 @@ const checkSystemAssets = async (
 		} else {
 			// not a valid number
 			// check if id is a valid token symbol of the system parachain chain
-			if (assetId === '') {
+			if (isRelayNativeAsset(registry, assetId)) {
 				return;
 			}
 
@@ -816,10 +811,25 @@ const checkParaToRelayAssetId = (assetId: string, registry: Registry, specName: 
  *
  * @param assetIds
  */
-export const checkAssetIdsLengthIsValid = (assetIds: string[]) => {
-	if (assetIds.length > MAX_ASSETS_FOR_TRANSFER) {
+export const checkAssetIdsLengthIsValid = (
+	assetIds: string[],
+	xcmPalletName: XcmPalletName,
+	assetTransferType: string | undefined,
+) => {
+	if (assetIds.length > MAX_ASSETS_FOR_TRANSFER && !assetTransferType) {
 		throw new BaseError(
 			`Maximum number of assets allowed for transfer is 2. Found ${assetIds.length} assetIds`,
+			BaseErrorsEnum.InvalidInput,
+		);
+	}
+
+	if (
+		assetIds.length > 1 &&
+		!assetTransferType &&
+		(xcmPalletName === XcmPalletName.polkadotXcm || xcmPalletName === XcmPalletName.xcmPallet)
+	) {
+		throw new BaseError(
+			`transferAssets transactions cannot contain more than 1 asset location id. Found ${assetIds.length} assetIds`,
 			BaseErrorsEnum.InvalidInput,
 		);
 	}
@@ -886,7 +896,7 @@ export const checkAssetIdsAreOfSameAssetIdType = (assetIds: string[]) => {
 			const isValidInt = validateNumber(assetId);
 			if (isValidInt) {
 				integerAssetIdFound = assetId;
-			} else if (assetId.toLowerCase().includes('parents')) {
+			} else if (assetIdIsLocation(assetId)) {
 				multiLocationAssetIdFound = assetId;
 			} else {
 				symbolAssetIdFound = assetId;
@@ -917,6 +927,66 @@ export const checkAssetIdsAreOfSameAssetIdType = (assetIds: string[]) => {
 };
 
 /**
+ * Checks to ensure that the xcmVersion is at least 3 for Bridge transactions
+ *
+ * @param xcmDirection
+ * @param xcmVersion
+ */
+export const checkXcmVersionIsValidForBridgeTx = (xcmVersion: number) => {
+	if (xcmVersion && xcmVersion < 3) {
+		throw new BaseError('Bridge transactions require XCM version 3 or greater', BaseErrorsEnum.InvalidXcmVersion);
+	}
+};
+
+/**
+ * Checks to ensure that required inputs are provided for Bridge transactions
+ *
+ * @param paysWithFeeDest
+ * @param assetTransferType
+ * @param remoteReserveAssetTransferTypeLocation
+ * @param feesTransferType
+ * @param remoteReserveFeesTransferTypeLocation
+ */
+export const checkBridgeTxInputs = (
+	paysWithFeeDest: string | undefined,
+	assetTransferType: string | undefined,
+	remoteReserveAssetTransferTypeLocation: string | undefined,
+	feesTransferType: string | undefined,
+	remoteReserveFeesTransferTypeLocation: string | undefined,
+) => {
+	if (assetTransferType && !paysWithFeeDest) {
+		throw new BaseError(
+			'paysWithFeeDest input is required for bridge transactions when assetTransferType is provided',
+			BaseErrorsEnum.InvalidInput,
+		);
+	}
+	if (assetTransferType && assetTransferType === 'RemoteReserve' && !remoteReserveAssetTransferTypeLocation) {
+		throw new BaseError(
+			'remoteReserveAssetTransferTypeLocation input is required for bridge transactions when asset transfer type is RemoteReserve',
+			BaseErrorsEnum.InvalidInput,
+		);
+	}
+	if (feesTransferType && feesTransferType === 'RemoteReserve' && !remoteReserveFeesTransferTypeLocation) {
+		throw new BaseError(
+			'remoteReserveFeeAssetTransferTypeLocation input is required for bridge transactions when fee asset transfer type is RemoteReserve',
+			BaseErrorsEnum.InvalidInput,
+		);
+	}
+};
+
+export const checkPaysWithFeeDestAssetIdIsInAssets = (assetIds: string[], paysWithFeeDest: string | undefined) => {
+	if (!paysWithFeeDest) {
+		return;
+	}
+
+	if (!assetIds.includes(paysWithFeeDest)) {
+		throw new BaseError(
+			`paysWithFeeDest asset must be present in assets to be transferred. Did not find ${paysWithFeeDest} in ${assetIds.toString()}`,
+		);
+	}
+};
+
+/**
  * Checks to ensure that the xcmVersion is at least 3 if paysWithFeeDest is provided
  *
  * @param xcmVersion
@@ -934,7 +1004,7 @@ export const checkXcmVersionIsValidForPaysWithFeeDest = (
 		xcmVersion &&
 		xcmVersion < 3
 	) {
-		throw new BaseError('paysWithFeeDest requires XCM version 3', BaseErrorsEnum.InvalidXcmVersion);
+		throw new BaseError('paysWithFeeDest requires XCM version 3 or greater', BaseErrorsEnum.InvalidXcmVersion);
 	}
 };
 
@@ -980,7 +1050,7 @@ export const checkAssetIdInput = async (
 		checkIfAssetIdIsBlankSpace(assetId);
 
 		if (xcmDirection === Direction.RelayToSystem) {
-			checkRelayToSystemAssetId(assetId, relayChainInfo);
+			checkRelayToSystemAssetId(assetId, relayChainInfo, registry);
 		}
 
 		if (xcmDirection === Direction.RelayToPara) {
@@ -988,7 +1058,7 @@ export const checkAssetIdInput = async (
 		}
 
 		if (xcmDirection === Direction.SystemToRelay) {
-			checkSystemToRelayAssetId(assetId, relayChainInfo);
+			checkSystemToRelayAssetId(assetId, relayChainInfo, registry);
 		}
 
 		if (xcmDirection === Direction.SystemToPara) {
@@ -1041,7 +1111,16 @@ export const checkAssetIdInput = async (
  */
 export const checkXcmTxInputs = async (baseArgs: XcmBaseArgsWithPallet, opts: CheckXcmTxInputsOpts) => {
 	const { api, direction, assetIds, amounts, destChainId, xcmVersion, specName, registry, xcmPallet } = baseArgs;
-	const { paysWithFeeDest, isForeignAssetsTransfer, isLiquidTokenTransfer, isPrimaryParachainNativeAsset } = opts;
+	const {
+		paysWithFeeDest,
+		isForeignAssetsTransfer,
+		isLiquidTokenTransfer,
+		isPrimaryParachainNativeAsset,
+		assetTransferType,
+		remoteReserveAssetTransferTypeLocation,
+		feesTransferType,
+		remoteReserveFeesTransferTypeLocation,
+	} = opts;
 	const relayChainInfo = registry.currentRelayRegistry;
 
 	if (isPrimaryParachainNativeAsset) {
@@ -1068,17 +1147,12 @@ export const checkXcmTxInputs = async (baseArgs: XcmBaseArgsWithPallet, opts: Ch
 	/**
 	 * Checks to ensure that assetId's have a length no greater than MAX_ASSETS_FOR_TRANSFER
 	 */
-	checkAssetIdsLengthIsValid(assetIds);
+	checkAssetIdsLengthIsValid(assetIds, xcmPallet, assetTransferType);
 
 	/**
 	 * Checks to ensure that assetId's have no duplicate values
 	 */
 	checkAssetIdsHaveNoDuplicates(assetIds);
-
-	/**
-	 * Checks to ensure that assetId's are either empty string, symbol and integer values or multilocations
-	 */
-	checkAssetIdsAreOfSameAssetIdType(assetIds);
 
 	/**
 	 * Checks to ensure that assetId's are either valid integer numbers or native asset token symbols
@@ -1105,6 +1179,21 @@ export const checkXcmTxInputs = async (baseArgs: XcmBaseArgsWithPallet, opts: Ch
 		checkRelayAmountsLength(amounts);
 	}
 
+	if (direction === Direction.RelayToBridge) {
+		checkRelayAssetIdLength(assetIds);
+		checkRelayAmountsLength(amounts);
+		getGlobalConsensusSystemName(destChainId);
+		checkBridgeTxInputs(
+			paysWithFeeDest,
+			assetTransferType,
+			remoteReserveAssetTransferTypeLocation,
+			feesTransferType,
+			remoteReserveFeesTransferTypeLocation,
+		);
+		checkPaysWithFeeDestAssetIdIsInAssets(assetIds, paysWithFeeDest);
+		checkXcmVersionIsValidForBridgeTx(xcmVersion);
+	}
+
 	if (direction === Direction.SystemToRelay) {
 		checkRelayAssetIdLength(assetIds);
 		checkRelayAmountsLength(amounts);
@@ -1115,7 +1204,6 @@ export const checkXcmTxInputs = async (baseArgs: XcmBaseArgsWithPallet, opts: Ch
 			checkMultiLocationIdLength(assetIds);
 			checkMultiLocationAmountsLength(amounts);
 			checkAssetsAmountMatch(assetIds, amounts);
-			checkAssetsAmountMatch(assetIds, amounts);
 			checkMultiLocationsContainOnlyNativeOrForeignAssetsOfDestChain(direction, destChainId, assetIds);
 		}
 		checkAssetsAmountMatch(assetIds, amounts);
@@ -1123,12 +1211,26 @@ export const checkXcmTxInputs = async (baseArgs: XcmBaseArgsWithPallet, opts: Ch
 
 	if (direction === Direction.SystemToSystem) {
 		if (isForeignAssetsTransfer) {
-			checkMultiLocationIdLength(assetIds);
-			checkMultiLocationAmountsLength(amounts);
 			checkAssetsAmountMatch(assetIds, amounts);
 			checkMultiLocationsContainOnlyNativeOrForeignAssetsOfDestChain(direction, destChainId, assetIds);
 		}
 		checkIfNativeRelayChainAssetPresentInMultiAssetIdList(assetIds, registry);
+	}
+
+	if (direction === Direction.SystemToBridge) {
+		checkMultiLocationIdLength(assetIds);
+		checkMultiLocationAmountsLength(amounts);
+		checkAssetsAmountMatch(assetIds, amounts);
+		getGlobalConsensusSystemName(destChainId);
+		checkBridgeTxInputs(
+			paysWithFeeDest,
+			assetTransferType,
+			remoteReserveAssetTransferTypeLocation,
+			feesTransferType,
+			remoteReserveFeesTransferTypeLocation,
+		);
+		checkPaysWithFeeDestAssetIdIsInAssets(assetIds, paysWithFeeDest);
+		checkXcmVersionIsValidForBridgeTx(xcmVersion);
 	}
 
 	if (direction === Direction.ParaToSystem || direction === Direction.ParaToPara) {